# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      error_trace:
        description: Full error + stack trace (non-prod only)
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorData'
      success:
        example: false
        type: boolean
    type: object
...
  upload.swaggerDeleteResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  upload.swaggerListUploadResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/upload.swaggerUploadData'
        type: array
      success:
        type: boolean
    type: object
  upload.swaggerUploadData:
    properties:
      created_at:
        type: string
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
  upload.swaggerUploadResponse:
    properties:
      data:
        $ref: '#/definitions/upload.swaggerUploadData'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: API server for booking system.
  title: PhotoStudio API
  version: "1.0"
paths:
...
  /uploads:
    get:
      description: Возвращает список всех файлов, загруженных текущим авторизованным
        пользователем.
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            $ref: '#/definitions/upload.swaggerListUploadResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мои файлы
      tags:
      - Uploads
    post:
      consumes:
      - multipart/form-data
      description: Принимает multipart/form-data с полем "file". Возвращает метаданные
        загруженного файла.
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Успешная загрузка
          schema:
            $ref: '#/definitions/upload.swaggerUploadResponse'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "413":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузка файла
      tags:
      - Uploads
  /uploads/{id}:
    delete:
      description: Удаляет файл и его метаданные. Доступно только владельцу файла.
      parameters:
      - description: ID файла
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/upload.swaggerDeleteResponse'
        "401":
          description: Ошибка прав доступа
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Ошибка прав доступа
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить файл
      tags:
      - Uploads
    get:
      description: Возвращает информацию о ранее загруженном файле по его идентификатору.
      parameters:
      - description: ID файла
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            $ref: '#/definitions/upload.swaggerUploadResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить метаданные файла
      tags:
      - Uploads
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"